fix: file path, naming, function calls, exception logic
Removed file from root directory, corrected naming from parent class to child class, removed unused imports, fixed function calls to match required parameters, updated exception messages

docs: added comments and function outlines
